// src/app/common/utils/app/producto/producto.interface.ts
export interface Producto {

    cod_producto: string;
    cod_hash: string;
    codigo: string;
    tipo: string;
    categoria: string;
    talla: string;
    color: string;
    material: string;
    sexo: string;
    descripcion: string;
    precio_unitario: number;
    precio_mayor: number;
    estado: number;

}